Fixes #10:
I'm moving up the line that checks if the event type is a message. This is because where it is now, it first tries to shortUserId with event.user.id and there is a possiblity that evet.user is undefined (e.g when we have botpress-webhooks installed and the webhook is triggered).